+++ src/tests/Physics/PHPenaltyTest/main.cpp
@@ -100,7 +100,7 @@ void SPR_CDECL display(){
 		CDShapeIf* shape = soFloor->GetShape(i);
 		CDConvexMeshIf* mesh = DCAST(CDConvexMeshIf, shape);
 		Vec3f* base = mesh->GetVertices();
-		for(size_t f=0; f<mesh->NFace();++f){
+		for(int f=0; f<mesh->NFace();++f){
 			CDFaceIf* face = mesh->GetFace(f);
 			
 			glBegin(GL_POLYGON);
@@ -126,7 +126,7 @@ void SPR_CDECL display(){
 			CDShapeIf* shape = soBlock[n]->GetShape(i);
 			CDConvexMeshIf* mesh = DCAST(CDConvexMeshIf, shape);
 			Vec3f* base = mesh->GetVertices();
-			for(size_t f=0; f<mesh->NFace();++f){
+			for(int f=0; f<mesh->NFace();++f){
 				CDFaceIf* face = mesh->GetFace(f);
 				
 				glBegin(GL_POLYGON);
